% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs_wrappers.R
\name{dvs_version}
\alias{dvs_version}
\title{DVS Version}
\source{
Generated by miniextendr from Rust fn \code{dvs_version}
}
\usage{
dvs_version()
}
\value{
A character string with the version.
}
\description{
Returns the version of the DVS Rust backend.
}
\examples{
dvs_version()
}
